Read Title + URL from Google Scholar
The program can read links and titles from Google Scholar page. It prints the output in console. It can filter the results and create a separate list with only links from Science Direct page. For now, the focus is to scrape data only from that website. No UI other than a single input asking for term to be searched.