Can now find people
Program can scrape science direct websites and get their names. It creates a search query in Scopus using the name. The affiliation is difficult to scrape as it works on JavaScript. The python module cannot seem to extract the <dl> tag as well. If it could our project is practically done.  Anyway, there is a better alternative, the Scrappy module. The results are compiled together and placed in an excel file. The search query needs to be debugged as well, as for characters with diacritical marks it does not form a valid query. Those alphabets have to be mapped with the proper escape sequence characters.