API Service
The API to communicate between the frontend and the backend has been created. The server side now has the docker container to remove itself after ending. The two pages in the react app can now pass data and the overall dynamic structures of the application are now done. Contexts have been used to transfer data as well.