Front-end Update
The front end results page has now been completely upgraded. The filters tab is now functional and the front-end can now communicate with the backend to provide search queries.

If the URL is invalid, a page for that has been created. If the results are empty an empty results is shown. The footer has been created. The options for the sources and other parts in the filtering menu is now in a JSON file. A number of smaller bug fixes have also been implemented.